[chore/#413] benchmark 빌드 타입 추가 및 build-logic 리팩토링
