+++ app/build.gradle.kts
@@ -56,22 +56,10 @@ android {
     }
 
     buildTypes {
-        debug {
+        getByName("debug") {
             applicationIdSuffix = ".debug"
-            buildConfigField(
-                "String",
-                "BASE_URL",
-                properties["dev.base.url"] as String
-            )
         }
-
-        release {
-            buildConfigField(
-                "String",
-                "BASE_URL",
-                properties["prod.base.url"] as String
-            )
-
+        getByName("release") {
             isMinifyEnabled = true
             isShrinkResources = true
             proguardFiles(
@@ -80,10 +68,10 @@ android {
             )
             signingConfig = signingConfigs.getByName("release")
         }
-    }
-
-    buildFeatures {
-        buildConfig = true
+        getByName("benchmark") {
+            isMinifyEnabled = false
+            isShrinkResources = false
+        }
     }
 }
