+++ build-logic/src/main/java/hilingual.android.application.gradle.kts
@@ -1,6 +1,7 @@
+import com.android.build.api.dsl.ApplicationExtension
+import com.hilingual.build_logic.configureBuildTypes
 import com.hilingual.build_logic.configureHiltAndroid
 import com.hilingual.build_logic.configureKotlinAndroid
-
 
 plugins {
     id("com.android.application")
@@ -8,3 +9,17 @@ plugins {
 
 configureKotlinAndroid()
 configureHiltAndroid()
+
+extensions.configure<ApplicationExtension> {
+    configureBuildTypes(this)
+
+    buildTypes.getByName("benchmark") {
+        signingConfig = signingConfigs.getByName("debug")
+    }
+
+    sourceSets {
+        getByName("benchmark") {
+            res.srcDirs("src/release/res")
+        }
+    }
+}
