+++ build-logic/src/main/java/hilingual.android.library.gradle.kts
@@ -1,6 +1,8 @@
-import com.hilingual.build_logic.configureKotlinAndroid
+import com.android.build.gradle.LibraryExtension
+import com.hilingual.build_logic.configureBuildTypes
 import com.hilingual.build_logic.configureCoroutineAndroid
 import com.hilingual.build_logic.configureHiltAndroid
+import com.hilingual.build_logic.configureKotlinAndroid
 
 plugins {
     id("com.android.library")
@@ -9,3 +11,8 @@ plugins {
 configureKotlinAndroid()
 configureCoroutineAndroid()
 configureHiltAndroid()
+
+
+extensions.configure<LibraryExtension> {
+    configureBuildTypes(this)
+}
